backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
billiard==4.2.0
black==24.8.0
celery==5.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.2.1
Django==5.1
django-ckeditor-5==0.2.18
django-cors-headers==4.4.0
django-extensions==3.2.3
django-redis==5.4.0
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.3.0
executing==2.2.0
flake8==7.1.1
idna==3.10
inflection==0.5.1
ipython==8.36.0
isort==5.13.2
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sorl-thumbnail==12.11.0
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
